Use an explicit typecase to avoid a compiler warninga about implicit shortening of a 64-bit datatype to a 32-bit datatype.

+++ contrib/mod_tls.c
@@ -1251,7 +1251,7 @@ static void tls_prepare_provider_fds(int stdout_fd, int stderr_fd) {
     nfiles = 255;
 
   } else
-    nfiles = rlim.rlim_max;
+    nfiles = (unsigned long) rlim.rlim_max;
 #else /* no RLIMIT_NOFILE or RLIMIT_OFILE */
    nfiles = 255;
 #endif
